Fix lists breaking when trying to parse an non-node
Having a card, for example, linked into the tree by an arrow would break 
the list creator because a card's data structure is different from that 
of a "shape" (what we use for most nodes). This change adds logic to 
make sure that parents and children are only added if they're valid 
nodes: "shape" objects with titles.

resolves #13